requirements.txt
##### `requirements.txt`
##### QR Code Scanner for HR Staff Portal
##### Open-Source, hostet on https://github.com/DrBenjamin/QRCodeScanner
##### Please reach out to ben@benbox.org for any questions
### All needed libraries (and their versions)
streamlit == 1.17.0
streamlit_image_select
streamlit_qrcode_scanner
streamlit_ws_localstorage
websockets
gsheetsdb
gspread
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-cloud==0.21.1
qrcode